Allow nbsp in p tags - proof of concept
This makes it so when you make new line breaks, the empty paragraph tags have `nbsp` inside of them, which makes them take up space on the page just like they do in the editor. Since this is just a proof of concept, it comments out the valid elements file and for now allows ONLY p tags.